pyproject.toml
[tool.poetry]
name = "wikimedia-bot"
version = "0.1.0"
description = ""
authors = ["Nathan Heller <nathanhellbhoy@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.34.154"
mastodon-py = "^1.8.1"
requests = "^2.32.3"
pillow = "^10.4.0"
poetry-plugin-export = "^1.8.0"
tweepy = "^4.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
